backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
redis==5.2.1
celery[redis]==5.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
httpx==0.28.1
pydantic[email]==2.10.3